Refactor models into separate module
