+++ deploy.sh
@@ -8,7 +8,7 @@ echo "📦 Instalando dependencias..."
 pip install -r requirements.txt
 
 echo "🗄️  Aplicando migraciones de base de datos..."
-export FLASK_APP=app.py
+export FLASK_APP=app
 rm -rf migrations || true
 flask db init
 flask db migrate -m "migración inicial completa"
